requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
numpy==1.19.1
packaging==20.9
Pillow==8.1.0
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scipy==1.5.4
six==1.15.0
tokenizers==0.10.1
tqdm==4.56.1
transformers==4.3.1
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.4.0
gunicorn==20.0.4
torch==1.3.1+cpu 
torchvision==0.4.2+cpu
